requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
typing_extensions==4.8.0
urllib3==2.1.0
websocket-client==1.6.4
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.3
